Restore fs123-trsum
A bad cut-and-paste when we libraryized the server resulted in no
fs123-trsum header in the replies sent by fs123p7exportd.

So we haven't been getting end-to-end checksum verification(!)

However, even if the HTTP reply lacks an fs123-trsum header, the
client has always computed the trsum of the body, serialized it in the
diskcache, and verified it when deserializing.  So we haven't been
verifying the integrity of network transport, but we have been
verifying the integrity of the cache.

Change-Id: I008b1cc74bf4d2290a5e736e47fcd64dcd064e28